# Allow symbol (i.e. quoted identifier) as safe SQL string